Fixed some of the runtime errors
Note: Previously it was not stuck in the while loop as we thought, it was actually just waiting for the input (e.g. "dummy.txt")

When I gave it the input however, it crashed due to memory releasing issues. I fixed these memory issues but now, however, it is in fact stuck in the while loop.